Complete Week 1 portfolio and introductory Kotlin exercises

+++ tasks/task3_2/src/Main.kt
@@ -8,6 +8,6 @@ fun main(args: Array<String>) {
         exitProcess(1)
     }
 
-    val number = args[0].toInt()
+    val number = args[0].toDouble()
     println(number * number)
 }
